requirements.txt
numpy
opencv-python
Pillow
opencv-contrib-python
